backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic[email]==2.4.2
email-validator==2.1.0.post1
stripe==7.6.0
google-generativeai==0.3.1
python-docx==1.0.1
reportlab==4.0.7
Pillow==10.1.0
requests==2.31.0
beautifulsoup4==4.12.2 